# docs/06_API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: Unify V9 - Shared Host Ready - API
  description: |
    API for Unify University Assistant - Pars Pack Cloud Host + Host Iran
    Stack: Laravel 10 PHP 8.2 + MySQL + Local Filesystem + Polling + Cron + Pushe PHP
    600 Students, No VPS, No iOS, Android APK optional
    All dates stored Gregorian DATETIME UTC, displayed Shamsi YYYY/MM/DD via Morilog\Jalali
    Idempotency via MySQL table IdempotencyKeys, key UUID v4 header Idempotency-Key
    Auth via Laravel Sanctum SPA httpOnly cookie + CSRF
    Rate limiting via Laravel throttle file cache
  version: 9.0.0
  contact:
    name: Unify Team
servers:
  - url: https://api.unify-cs.ac.ir/api/v1
    description: Production API on Cloud Host (api subdomain)
  - url: http://localhost:8000/api/v1
    description: Local dev
security:
  - Sanctum: []
paths:
...
  /resources:
    get:
      summary: List evergreen resources filtered course+prof, paginated 20, sort newest/oldest/rated/downloaded
      parameters:
        - {name: course_id, in: query, schema: {type: string}}
        - {name: professor_id, in: query, schema: {type: string}}
        - {name: specification_id, in: query, schema: {type: string}}
        - {name: sort, in: query, schema: {type: string, enum: [newest, oldest, rated, downloaded], default: newest}}
        - {name: search, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer}}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: {type: array, items: {$ref: '#/components/schemas/Resource'}}
  /resources/{id}:
    get:
      summary: Get resource detail
      parameters: [{name: id, in: path, required: true, schema: {type: string}}]
      responses:
        '200': {content: {application/json: {schema: {$ref: '#/components/schemas/Resource'}}}}
        '404': {description: Not found or hard deleted}
...
components:
  securitySchemes:
    Sanctum:
      type: apiKey
      in: cookie
      name: laravel_session
  schemas:
...
    Resource:
      type: object
      properties:
        id: {type: string}
        course_id: {type: string}
        professor_id: {type: string}
        title: {type: string}
        description: {type: string}
        file_size_bytes: {type: integer}
        file_mime: {type: string}
        shamsi_original: {type: string}
        created_at_g: {type: string, format: date-time}
        status: {type: string, enum: [pending, approved, rejected]}
        version: {type: integer}
        family_id: {type: string}
        average_rating: {type: number}
        rating_count: {type: integer}
        download_count: {type: integer}
        badge_type: {type: string, enum: [professor, expert_approved, admin_approved], nullable: true}
        is_superseded: {type: boolean}
...